SCOTT: WIP disk-based graphics update

+++ terps/scott/detectgame.h
@@ -13,7 +13,7 @@
 #include <stdio.h>
 
 GameIDType DetectGame(const char *file_name);
-int SeekIfNeeded(int expected_start, int *offset, uint8_t **ptr);
+int SeekIfNeeded(int expected_start, size_t *offset, uint8_t **ptr);
 int TryLoading(struct GameInfo info, int dict_start, int loud);
 DictionaryType GetId(size_t *offset);
 int FindCode(const char *x, int base);
